* changes:
  Reorganize some code and fix checkstyle
  Rework the SONAR/Checkstyle report
  Fix a potential NPE
  Add Request ID in Common Logging
  Add 2 new methods getNodeValueByName
  Add missing the eventName
  Rework the javadoc
  Rework the format of this file
  FIx the POM for HTTPS param and jks file in test
  Fix the clds-reference json

+++ src/main/java/org/onap/clamp/clds/client/OperationalPolicyDeleteDelegate.java
@@ -56,13 +56,15 @@ public class OperationalPolicyDeleteDelegate implements JavaDelegate {
         prop.setCurrentModelElementId(policy.getId());
 
         String responseMessage = "";
-        for (PolicyChain policyChain : policy.getPolicyChains()) {
-            prop.setPolicyUniqueId(policyChain.getPolicyId());
-            responseMessage = policyClient.deleteBrms(prop);
-        }
-        if (responseMessage != null) {
-            execution.setVariable("operationalPolicyDeleteResponseMessage", responseMessage.getBytes());
-        }
+		if(policy.isFound()){
+	        for (PolicyChain policyChain : policy.getPolicyChains()) {
+	            prop.setPolicyUniqueId(policyChain.getPolicyId());
+	            responseMessage = policyClient.deleteBrms(prop);
+	        }
+	        if (responseMessage != null) {
+	            execution.setVariable("operationalPolicyDeleteResponseMessage", responseMessage.getBytes());
+	        }
+	    }
     }
 
 }
